requirements.txt
langchain
langchain-community
langchain-core
langchain-google-genai
pypdf
aiofiles
fastapi
uvicorn
jinja2
python-multipart
PyPDF2
faiss-cpu
python-dotenv
-e .